requirements.txt
arrow==1.3.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-redis==5.4.0
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
times==0.7
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
